Rename modularChangeEncoding -> modularChangeCodecs
